feat: Allow setting a lambda version description from S3 Metadata
Uses the S3 Metadata "description" during a deployment from S3 to set
the new created function version description to its value. This allows
to supply some metadata, e.g. the git commit hash when deploying via
S3.

The process creating the S3 object needs to set this metadata in order
for this to work.

If the metadatat is not set, the code uses the function description
for the version description by default.

Sadly `boto3` does not expose setting the version description in one
step with `update_function_code` with `Publish=True`, so two calls are
necessary for setting the description.

Note that no new IAM rights are necessary for `head_object`,
CodePipeline already has `Get*`, and in some AWS quirk `GetObject`
grants `HeadObject`.