fix(deployment): Wait for code update to complete before publish
In the branch where the version get's released after the code is
published a waiter was missing. This lead to errors like

```
botocore.errorfactory.ResourceConflictException:
An error occurred (ResourceConflictException) when calling the
PublishVersion operation: The operation cannot be performed at this
time. An update is in progress for resource:
arn:aws:lambda:eu-central-1:xxx:function:yyy